Move most rendering and setup code to `DecorationAdapter`
