+++ src/server/shell/decoration/basic_manager.cpp
@@ -96,19 +96,21 @@ void msd::BasicManager::decorate(std::shared_ptr<ms::Surface> const& surface)
         auto decoration = decoration_builder(locked_shell, surface);
         lock.lock();
         decoration->set_scale(scale);
-        decorations[surface.get()] = std::move(decoration);
+        decorations[surface.get()] = decoration;
     }
 }
 
 void msd::BasicManager::undecorate(std::shared_ptr<ms::Surface> const& surface)
 {
-    std::unique_ptr<Decoration> decoration;
+    // TODO: probably only frees one pointer, but not the whole set of things
+    // pointing at the decoration.
+    std::shared_ptr<Decoration> decoration;
     {
         std::lock_guard lock{mutex};
         auto const it = decorations.find(surface.get());
         if (it != decorations.end())
         {
-            decoration = std::move(it->second);
+            decoration = it->second;
             decorations.erase(it);
         }
     }
@@ -118,7 +120,7 @@ void msd::BasicManager::undecorate(std::shared_ptr<ms::Surface> const& surface)
 
 void msd::BasicManager::undecorate_all()
 {
-    std::vector<std::unique_ptr<Decoration>> to_destroy;
+    std::vector<std::shared_ptr<Decoration>> to_destroy;
     {
         std::lock_guard lock{mutex};
         for (auto& it : decorations)
